Add files via upload
The code is of a Library Management System in Python using object-oriented programming (OOP). It consists of two classes, Library and Student, with methods that allow borrowing and returning books.
LOGIC:
1. Student class:
-Methods:
request_book =  Prompts the user to enter the name of the book they want to borrow and returns the name.
return_book = Prompts the user to enter the name of the book they want to return and returns the name.

2. Library class:
-Methods:
display_available_books = Prints all the books present in library.
borrow_book = a. Checks if requested book is in the library
                          b. if yes , removes the book from the library
                          c. If no, says book is not available.
return_book: Adds the returned book to library

3. Main Function:
a. Creates an instance of Library with list of books
b. Creates an instance of Student to perform student interactions.
c.  Creates a while loop until the user exits.